Compute median ground elevation in grid cell in EptInPolygons
Read the ground points (class 2). Only test if the point is within the grid
index cell, but do not do a pip for the footprint itself.
When all ground point are read for the grid, compute the median Z per cell.
Output the median ground elevation for each polygon in the ground_heights output.